Find candidate

This is an unexported helper for acronym. The function is used wrapped in a tryCatch() that uses withTimeout to manage maximum wait time for the candidate acronym search.

find_candidate(collapsed, acronym_length, probs, dictionary, words_len)

Arguments

collapsed
The collapsed string of characters generated by mince

acronym_length
Number of characters in acronym; default is 3

probs
Vector of probabilities for selecting each character while generating candidate

dictionary
Character vector containing dictionary of terms from which acronym should be created; default is NULL and hunspell "en_us" dictionary will be used

words_len
Vector of the length of each word in the input

Value

Named list with three elements:

formatted: The candidate acronym and string with letters used capitalized

prefix: The candidate acronym

suffix: Words used with letters in acronym capitalized #'